requirements.txt
speechrecognition
gTTS
pyaudio
pydub
requests
python-dotenv
loguru
pydantic-settings
sounddevice
pvporcupine